+++ MonoZelda/Program.cs
@@ -1,0 +1,3 @@
+﻿
+using var game = new PixelPushers.MonoZelda.MonoZeldaGame();
+game.Run();
