+++ MonoZelda/Sprites/ISprite.cs
@@ -6,6 +6,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
+namespace PixelPushers.MonoZelda;
+
 interface ISprite
 {
